# strkey: reduce allocs in Encode