--- a/strkey/internal/crc16/main.go
+++ b/strkey/internal/crc16/main.go
@@ -47,8 +47,6 @@
 package crc16
 
 import (
-	"bytes"
-	"encoding/binary"
 	"errors"
 )
 
@@ -92,26 +90,21 @@ var crc16tab = [256]uint16{
 }
 
 // Checksum returns the 2-byte checksum for the provided data
-func Checksum(data []byte) []byte {
+func Checksum(data []byte) uint16 {
 	var crc uint16
-	var out [2]byte
 	for _, b := range data {
 		crc = ((crc << 8) & 0xffff) ^ crc16tab[((crc>>8)^uint16(b))&0x00FF]
 	}
-
-	binary.LittleEndian.PutUint16(out[:], crc)
-
-	return out[:]
+	return crc
 }
 
 // Validate returns an error if the provided checksum does not match
 // the calculated checksum of the provided data
-func Validate(data []byte, expected []byte) error {
-
+func Validate(data []byte, expected uint16) error {
 	actual := Checksum(data)
 
 	// validate the provided checksum against the calculated
-	if !bytes.Equal(actual, expected) {
+	if actual != expected {
 		return ErrInvalidChecksum
 	}