… strkey/encode_test.go
… strkey/internal/crc16/main_test.go