@leighmcculloch 2021-11-30
I originally opened this as an example of an idea I wanted to share on #4105, but in an attempt to validate the idea, I think this is essentially complete and could be merged. What do folks think?