@2opremio 2021-12-01
@tamirms do you think we should include this for the release? (it will reduce allocations in the path search algorithm).

If so, @leighmcculloch , could you change the base branch to `release-horizon-v2.12.0`. 

Thanks!