@leighmcculloch 2021-12-01
> @tamirms do you think we should include this for the release? (it will reduce allocations in the path search algorithm).
> 
> If so, @leighmcculloch , could you change the base branch to `release-horizon-v2.12.0`?

I've pushed this to master. Can you cherry-pick to the release if it is needed there?